詳細にアウトプットするクラスを分離
